Tag the correct commit
This works around a current limitation in `git-version-lite`, which has a hard-coded condition to tag the head of the incoming branch whenever the workflow is triggered by a PR even if that PR is now merged.

I argue that `git-version-lite` probably should not push tags at all; probably we should use `create-release` to do it. But if it's going to, it needs to accept—maybe require—a committish parameter.

This PR preserves the existing behavior of creating only a tag and not a GitHub release. I could be convinced we should create releases here instead.